Update README.md
Elaborated a bit on usage to make it more clear that by default install-subset is focused on devDependencies. And, how to utilize this fantastic tool for situations where you only want a subset of the application dependencies to be installed. In a docker container is a prime example, where I do not want nativescript and serverless installed as these are more specific to my machine or for purposes of deploying lambda functions via our CI/CD pipeline.